backend/requirements.txt
flask
flask-cors
torch
torchvision
numpy
Pillow
opencv-python-headless
matplotlib
gunicorn
